+++ landingzones/views.py
@@ -540,7 +540,9 @@ class ZoneMoveMixin(ZoneConfigPluginMixin):
         if validate_only:
             flow_data['validate_only'] = True
         else:
-            flow_data['move_verify'] = settings.LANDINGZONES_ZONE_MOVE_VERIFY
+            flow_data['move_verify'] = app_settings.get(
+                APP_NAME, 'zone_move_verify_site'
+            )
         taskflow.submit(
             project=project,
             user=request.user if request else None,
@@ -571,7 +573,7 @@ class ZoneResetMixin(ZoneConfigPluginMixin):
             task_found = any(
                 t.get('name').endswith('.submit_flow_task')
                 and 'args' in t
-                and t['args'][2].get('zone_uuid') == str(zone.sodar_uuid)
+                and t['args'][3].get('zone_uuid') == str(zone.sodar_uuid)
                 for t in v
             )
             if task_found: